Removes rustc_serialize dependency Adds implementations to convert to/from strings Adds category to Cargo.tom Moves convert test to its own file
